* fixed some cast warnings
git-svn-id: https://svn.code.sf.net/p/bzflag/code/trunk/bzflag@19256 08b3d480-bf2c-0410-a26f-811ee3361c24

+++ src/bzflag/bzflag.cxx
@@ -497,7 +497,7 @@ static void parse(int argc, char **argv)
           }
           c++;
         }
-        debugLevel += ((c - argv[i]) - 1);
+        debugLevel += (int)((c - argv[i]) - 1);
       }
     }
     else {
